Move docs svc to controller package, improve tests
