+++ components/helm-broker/internal/controller/clusteraddonsconfiguration_controller.go
@@ -183,8 +183,7 @@ func (r *ReconcileClusterAddonsConfiguration) addAddonsProcess(addon *addonsv1al
 
 	r.log.Info("- update AddonsConfiguration status")
 	r.statusSnapshot(addon, repositories)
-	err = r.updateAddonStatus(addon)
-	if err != nil {
+	if err = r.updateAddonStatus(addon); err != nil {
 		r.log.Errorf("cannot update ClusterAddonsConfiguration %s: %v", addon.Name, err)
 		return exerr.Wrap(err, "while update AddonsConfiguration status")
 	}
