+++ components/helm-broker/internal/controller/docs_provider.go
@@ -1,4 +1,4 @@
-package bundle
+package controller
 
 import (
 	"context"
